+++ config/config.exs
@@ -16,7 +16,7 @@ config :macroy, MacroyWeb.Endpoint,
   secret_key_base: "VnPtEPAJpgFi+Sx5d9lFSLLaJHpuocfYA6BBtsTHqUMT6xl4grqIfNjFj8RHJFzE",
   render_errors: [view: MacroyWeb.ErrorView, accepts: ~w(html json)],
   pubsub: [name: MacroyWeb.PubSub, adapter: Phoenix.PubSub.PG2],
-  liveview: [signing_salt: System.get_env("secret_phxlview_key")]
+  live_view: [signing_salt: System.get_env("secret_phxlview_key")]
 
 # Configures Elixir's Logger
 config :logger, :console,