+++ lib/macroy_web/live/todo_live.ex
@@ -1,0 +1,10 @@
+defmodule MacroyWeb.TodoLive do
+  use Phoenix.LiveView
+  alias MacroyWeb.TodoView
+
+  def mount(%{todos: todos}, socket) do
+    {:ok, assign(socket, todos: todos)}
+  end
+
+  def render(assigns), do: TodoView.render("index.html", assigns)
+end