+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Installed Phoenix LiveView.
 
 ## [1.3.0] - 2019-11-30
 ### Added