+++ CHANGELOG.md
@@ -6,11 +6,16 @@ and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.3.1] - 2019-12-01
 ### Added
 - Installed Phoenix LiveView.
 
+### Changed
+- Converted Todo index to a Phoenix Live View.
+
 ### Fixed
 - Spelling error with the Phoenix Live View signing salt.
+
 
 ## [1.3.0] - 2019-11-30
 ### Added
@@ -143,7 +148,7 @@ work, and can be closed.
 - Link from OrgFile index to home.
 - Link from home to OrgFile index.
 
-[Unreleased]: https://github.com/brotherjack/macroy/compare/v1.3.0...HEAD
+[Unreleased]: https://github.com/brotherjack/macroy/compare/v1.3.1...HEAD
 [0.0.1]: https://github.com/brotherjack/macroy/releases/tag/v0.0.1
 [0.0.2]: https://github.com/brotherjack/macroy/compare/v0.0.1...v0.0.2
 [0.1.0]: https://github.com/brotherjack/macroy/compare/v0.0.2...v0.1.0
@@ -154,3 +159,4 @@ work, and can be closed.
 [1.2.2]: https://github.com/brotherjack/macroy/compare/v1.2.1...v1.2.2
 [1.2.3]: https://github.com/brotherjack/macroy/compare/v1.2.2...v1.2.3
 [1.3.0]: https://github.com/brotherjack/macroy/compare/v1.2.3...v1.3.0
+[1.3.1]: https://github.com/brotherjack/macroy/compare/v1.3.0...v1.3.1